tests/wpt/tests/css/css-pseudo/highlight-cascade/highlight-pseudos-visited-computed-001.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Pseudo-Elements Test: highlight selectors getComputedStyle on visited</title>
<link rel="help" href="https://drafts.csswg.org/css-pseudo/#highlight-selectors">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  a::selection {
    color: lime;
  }
  a::target-text {
    color: lime;
  }
  a::search-text {
    color: lime;
  }
  a::search-text:current {
    color: red;
  }
  a::spelling-error {
    color: lime;
  }
  a::grammar-error {
    color: lime;
  }
  a::highlight(foo) {
    color: lime;
  }
  a:visited::selection {
    color: yellow;
  }
  a:visited::target-text {
    color: yellow;
  }
  a:visited::search-text {
    color: yellow;
  }
  a:visited::search-text:current {
    color: red;
  }
  a:visited::spelling-error {
    color: yellow;
  }
  a:visited::grammar-error {
    color: yellow;
  }
  a:visited::highlight(foo) {
    color: yellow;
  }
</style>
<a id="target1" class="target" href=""></a>
<a id="target2" class="target" href="unvisited"></a>
<script>
  for (const pseudo of ["::selection", "::target-text", "::search-text", "::spelling-error", "::grammar-error", "::highlight(foo)"]) {
    for (const target of [target1, target2]) {
      test(() => {
        let style = getComputedStyle(target, pseudo);
        assert_equals(style.color, "rgb(0, 255, 0)", "Color is lime.");
      }, `getComputedStyle() for ${pseudo} at #${target.id}`);
    }
  }
</script>
